refactor(app): migrate findLastAt util to helpers.ts
